+++ src/providers.rs
@@ -5,7 +5,13 @@ mod mock;
 #[cfg(feature = "provider-openai")]
 mod openai;
 
+#[cfg(feature = "provider-ollama")]
+mod ollama;
+
 pub use mock::MockProvider;
 
 #[cfg(feature = "provider-openai")]
 pub use openai::{OpenAiProvider, OpenAiProviderConfig};
+
+#[cfg(feature = "provider-ollama")]
+pub use ollama::{OllamaProvider, OllamaProviderConfig};